SNAPSHOT from yday, still fails, incorrect LSN and probably the return value

+++ pgxn/neon/communicator_new.h
@@ -38,7 +38,7 @@ extern void communicator_new_prefetch_register_bufferv(NRelFileInfo rinfo, ForkN
 													   BlockNumber nblocks);
 extern bool communicator_new_cache_contains(NRelFileInfo rinfo, ForkNumber forkNum,
 											BlockNumber blockno);
-extern int	communicator_new_read_slru_segment(SlruKind kind, int64 segno,
+extern int	communicator_new_read_slru_segment(SlruKind kind, uint32_t segno,
 											   void *buffer);
 
 /* Write requests, to keep the caches up-to-date */
